Make skin color change persist through save object

+++ SaveObject.cs
@@ -8,9 +8,9 @@ public class SaveObject
     public bool leaf;
     public bool band;
     public bool stairDoorOpen;
-    public int red;
-    public int blue;
-    public int green;
+    public float red;
+    public float blue;
+    public float green;
 
 
 }
